Extract CLI specific code to a separate file
